third_party/WebKit/LayoutTests/webaudio/AudioNode/audionode-disconnect-audioparam-expected.txt

Test disconnect() method on AudioParam destination.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Channel #0 contains all the expected values in the correct order: [2.25,1.5].
PASS The index of value change is equal to 11008.
PASS Channel #0 contains all the expected values in the correct order: [3,1.5].
PASS The index of value change in channel #0 is equal to 11008.
PASS Channel #1 contains all the expected values in the correct order: [6,3].
PASS The index of value change in channel #1 is equal to 11008.
PASS gain1.disconnect(gain3.gain) threw InvalidAccessError: Failed to execute 'disconnect' on 'AudioNode': the given AudioParam is not connected..
PASS splitter.disconnect(gain1.gain, 1) threw InvalidAccessError: Failed to execute 'disconnect' on 'AudioNode': specified destination AudioParam and node output (1) are not connected..
PASS splitter.disconnect(gain1.gain, 2) threw IndexSizeError: Failed to execute 'disconnect' on 'AudioNode': The output index provided (2) is outside the range [0, 1]..
PASS successfullyParsed is true

TEST COMPLETE
